requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.1.0
cycler==0.11.0
datasets==2.12.0
dill==0.3.6
filelock==3.12.0
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.1
invisible-watermark
Jinja2==3.1.2
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch==2.20.0
opencv-python
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.0
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytz==2023.3
PyWavelets
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
responses==0.18.0
safetensors==0.3.1
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
